Changes made based on comments
